Add provider pattern to access new session info aware driver (#16242)
[AB#4765](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/4765)

## Description
In https://github.com/microsoft/FluidFramework/pull/16160 we added the
ability to surface session info from odsp-driver.

The host application developers prefer the component not to have a
dependency directly on the odsp-driver package. This makes sense as the
component developer should not be aware of the specific driver being
used.

This PR adds a provider so that the component receiving a concrete
instance of DocumentServiceFactory can discover if that instance has the
`getRelayServiceSessionInfo` function and be able to call it if it
exists.

## Example Usage

```
// Get DocumentServiceFactory object and ResolvedUrl object from dependency provider
const documentServiceFactory: IDocumentSeriviceFactory = await provider.getDocumentServiceFactory();
const resolvedUrl: IResolvedUrl = await provider.getResolvedUrl();

// Test for new interface that has the method that returns session info
const maybeSessionAwareDriverFactory: FluidObject<IProvideSessionAwareDriverFactory> = documentServiceFactory;
if (maybeSessionAwareDriverFactory?.ISessionAwareDriverFactory !== undefined) {
    // It implements the new interface. Call it and use the result
    const sessionInfo = maybeSessionAwareDriverFactory?.ISessionAwareDriverFactory.getRelayServiceSessionInfo(resolvedUrl);
    // Pass info to CNS
    } else { 
    // documentServiceFactory doesn't implement the new interface, so we're probably running on a host that doesn't use ODSP 
    }
}
```

---------
