+++ api-report/odsp-driver-definitions.api.md
@@ -7,6 +7,7 @@
 import { DriverError } from '@fluidframework/driver-definitions';
 import { IDriverErrorBase } from '@fluidframework/driver-definitions';
 import { IFluidResolvedUrl } from '@fluidframework/driver-definitions';
+import { IResolvedUrl } from '@fluidframework/driver-definitions';
 
 // @public (undocumented)
 export type CacheContentType = "snapshot" | "ops";
@@ -141,6 +142,18 @@ export interface IPersistedCache {
 }
 
 // @public
+export interface IProvideSessionAwareDriverFactory {
+    // (undocumented)
+    readonly IRelaySessionAwareDriverFactory: IRelaySessionAwareDriverFactory;
+}
+
+// @public
+export interface IRelaySessionAwareDriverFactory extends IProvideSessionAwareDriverFactory {
+    // (undocumented)
+    getRelayServiceSessionInfo(resolvedUrl: IResolvedUrl): Promise<ISocketStorageDiscovery | undefined>;
+}
+
+// @public
 export interface ISharingLink extends ISharingLinkKind {
     // (undocumented)
     webUrl: string;
@@ -166,6 +179,23 @@ export interface ISnapshotOptions {
     mds?: number;
     // (undocumented)
     timeout?: number;
+}
+
+// @public
+export interface ISocketStorageDiscovery {
+    // (undocumented)
+    deltaStorageUrl: string;
+    deltaStreamSocketUrl: string;
+    // (undocumented)
+    id: string;
+    refreshSessionDurationSeconds?: number;
+    // (undocumented)
+    runtimeTenantId?: string;
+    // (undocumented)
+    snapshotStorageUrl: string;
+    socketToken?: string;
+    // (undocumented)
+    tenantId: string;
 }
 
 // @public
