+++ packages/drivers/odsp-driver/package.json
@@ -107,6 +107,20 @@
 		"typescript": "~4.5.5"
 	},
 	"typeValidation": {
-		"broken": {}
+		"broken": {
+			"ClassDeclaration_OdspDocumentServiceFactoryWithCodeSplit": {
+				"forwardCompat": false
+			},
+			"ClassDeclaration_OdspDocumentServiceFactory": {
+				"forwardCompat": false
+			},
+			"ClassDeclaration_OdspDocumentServiceFactoryCore": {
+				"forwardCompat": false
+			},
+			"RemovedInterfaceDeclaration_ISocketStorageDiscovery": {
+				"forwardCompat": false,
+				"backCompat": false
+			}
+		}
 	}
 }
