+++ packages/drivers/odsp-driver/src/test/types/validateOdspDriverPrevious.generated.ts
@@ -112,26 +112,14 @@ use_old_InterfaceDeclaration_ISnapshotContents(
 /*
 * Validate forward compat by using old type in place of current type
 * If breaking change required, add in package.json under typeValidation.broken:
-* "InterfaceDeclaration_ISocketStorageDiscovery": {"forwardCompat": false}
+* "RemovedInterfaceDeclaration_ISocketStorageDiscovery": {"forwardCompat": false}
 */
-declare function get_old_InterfaceDeclaration_ISocketStorageDiscovery():
-    TypeOnly<old.ISocketStorageDiscovery>;
-declare function use_current_InterfaceDeclaration_ISocketStorageDiscovery(
-    use: TypeOnly<current.ISocketStorageDiscovery>);
-use_current_InterfaceDeclaration_ISocketStorageDiscovery(
-    get_old_InterfaceDeclaration_ISocketStorageDiscovery());
 
 /*
 * Validate back compat by using current type in place of old type
 * If breaking change required, add in package.json under typeValidation.broken:
-* "InterfaceDeclaration_ISocketStorageDiscovery": {"backCompat": false}
+* "RemovedInterfaceDeclaration_ISocketStorageDiscovery": {"backCompat": false}
 */
-declare function get_current_InterfaceDeclaration_ISocketStorageDiscovery():
-    TypeOnly<current.ISocketStorageDiscovery>;
-declare function use_old_InterfaceDeclaration_ISocketStorageDiscovery(
-    use: TypeOnly<old.ISocketStorageDiscovery>);
-use_old_InterfaceDeclaration_ISocketStorageDiscovery(
-    get_current_InterfaceDeclaration_ISocketStorageDiscovery());
 
 /*
 * Validate forward compat by using old type in place of current type
@@ -191,6 +179,7 @@ declare function get_old_ClassDeclaration_OdspDocumentServiceFactory():
 declare function use_current_ClassDeclaration_OdspDocumentServiceFactory(
     use: TypeOnly<current.OdspDocumentServiceFactory>);
 use_current_ClassDeclaration_OdspDocumentServiceFactory(
+    // @ts-expect-error compatibility expected to be broken
     get_old_ClassDeclaration_OdspDocumentServiceFactory());
 
 /*
@@ -215,6 +204,7 @@ declare function get_old_ClassDeclaration_OdspDocumentServiceFactoryCore():
 declare function use_current_ClassDeclaration_OdspDocumentServiceFactoryCore(
     use: TypeOnly<current.OdspDocumentServiceFactoryCore>);
 use_current_ClassDeclaration_OdspDocumentServiceFactoryCore(
+    // @ts-expect-error compatibility expected to be broken
     get_old_ClassDeclaration_OdspDocumentServiceFactoryCore());
 
 /*
@@ -239,6 +229,7 @@ declare function get_old_ClassDeclaration_OdspDocumentServiceFactoryWithCodeSpli
 declare function use_current_ClassDeclaration_OdspDocumentServiceFactoryWithCodeSplit(
     use: TypeOnly<current.OdspDocumentServiceFactoryWithCodeSplit>);
 use_current_ClassDeclaration_OdspDocumentServiceFactoryWithCodeSplit(
+    // @ts-expect-error compatibility expected to be broken
     get_old_ClassDeclaration_OdspDocumentServiceFactoryWithCodeSplit());
 
 /*
